+++ internal/gpu/sdf_render.go
@@ -299,7 +299,7 @@ func (p *SDFRenderPipeline) createPipeline() error {
 			{
 				Binding:    0,
 				Visibility: gputypes.ShaderStageVertex | gputypes.ShaderStageFragment,
-				Buffer:     &gputypes.BufferBindingLayout{Type: gputypes.BufferBindingTypeUniform},
+				Buffer:     &gputypes.BufferBindingLayout{Type: gputypes.BufferBindingTypeUniform, MinBindingSize: sdfRenderUniformSize},
 			},
 		},
 	})